requirements.txt
Django==6.0.3
django-mysql==4.19.0
mysqlclient==2.2.8
Pillow==12.1.1
python-decouple==3.8
